requirements.txt
streamlit
torch                   #==2.2.2
torchvision             #==0.17.2
tensorflow              #==2.15.0
efficientnet-pytorch    #==0.7.1
mediapipe               #==0.10.9
Pillow                  #==10.3.0
numpy                   #==1.26.4
